Big bug Fix for phi and dp cuts in Xbj_New_Fit.C
In the main elastic electron counting code the phi and dp cuts on Xbj were defined wrong when read in from the root file. They are now correct, but fits and cuts now need to be carefully reexamined. Also the dp cut on the Al background was defined incorrectly and is now fixed. 

High_Background.C was a code that was attempting to locate where the anomalous background after the elastic peak was coming from. This code let me realize the error in the main electron counting code and correct it. Once the errors were corrected the background after the elastic peak almost disappeared and I now think I can safely ignore most of it.