Single Density Profile Added Back
Added back in a single density profile as an option along with the split density profile. Also fixed a bug where the peak matching function for the yield didn't work when the SIMC results needed to be scaled up instead of down. The error was basically that I forgot to swap the - for a + when changing the scale factor upwards.