Better Fit Ranges and Parameters + Lorentzian Fit
Fixed the two exponential parameters in the total fits to the exponential only fits. The range of the exponential fits was also narrowed and the SIMC exponential makes sure to focus on fitting background with very few elastics under it. These changes pulled up the total fit to match data better and decreased the yield accordingly. Now both data and SIMC fits are much closer to agreement. SIMC is still too spread. Also added the option to fit with a Lorentzian instead of a Gsussian, but the results were disappointing. Next up is to add a Gaussian area matching function like the peak matching function so we are properly comparing yields as opposed to just peak heights.