First Iteration of Gaussian Area Matching
Added to the matching functions one that matches the areas of the Gaussian portions of the total data and total SIMC fits. This looks like it will significantly change our XS (I believe by decreasing it which puts us in better agreement with other measurements). Still need to test how robust this is as well as use it to rescale the XS in SIMC.